Update README.md
This case study provides a comprehensive analysis of car sales data in Abuja, Nigeria. It focused on various aspects such as the impact of vehicle age on sales prices, the popularity of car makes and models, the influence of engine type and color on sales prices, and the relationship between listing and final sale prices. The insights gained from this analysis are intended to help car dealerships and manufacturers better understand market dynamics and optimize their sales strategies.